fix(types): replace global.d.ts with explicit module exports
