+++ packages/constellation/src/components/Base/Typography/Text/Text.tsx
@@ -1,6 +1,7 @@
 import cx from 'classnames'
 import React, { forwardRef } from 'react'
 
+import { PolymorphicRef } from '../../../../typings/polymorphic-ref'
 import { SemanticColors } from '../../Colors'
 import { getTextColorClass } from '../util'
 import { TextComponentType } from '.'
