+++ packages/constellation/src/components/Base/Typography/Text/Text.types.ts
@@ -1,3 +1,4 @@
+import { PolymorphicComponentPropsWithRef } from '../../../../typings/polymorphic-component-props-with-ref'
 import { SemanticColors } from '../../Colors'
 
 export const textVariants = [
