+++ packages/constellation/src/components/Compositions/Alerts/Banner/Banner.tsx
@@ -1,6 +1,7 @@
 import cx from 'classnames'
 import React, { forwardRef } from 'react'
 
+import { PolymorphicRef } from '../../../../typings/polymorphic-ref'
 import { SemanticColors, Text } from '../../../Base'
 import { AlertBannerComponentType, AlertBannerProps } from '.'
