+++ packages/constellation/src/components/Compositions/Headers/Balance Header/BalanceHeader.tsx
@@ -1,6 +1,7 @@
 import cx from 'classnames'
 import React, { forwardRef } from 'react'
 
+import { PolymorphicRef } from '../../../../typings/polymorphic-ref'
 import { centsToDollarString } from '../../../../utils'
 import { SemanticColors, Text } from '../../../Base'
 import { IconButton } from '../../../Primitives'
