+++ packages/constellation/src/components/Compositions/Headers/Section Header/SectionHeader.tsx
@@ -1,6 +1,7 @@
 import cx from 'classnames'
 import React, { forwardRef } from 'react'
 
+import { PolymorphicRef } from '../../../../typings/polymorphic-ref'
 import { Component as ComponentType, Props } from './SectionHeader.types'
 
 /**
