+++ packages/constellation/src/components/Primitives/Buttons/Link/Link.tsx
@@ -1,6 +1,7 @@
 import cx from 'classnames'
 import React, { forwardRef } from 'react'
 
+import { PolymorphicRef } from '../../../../typings/polymorphic-ref'
 import { IconCheck } from '../../../Base'
 import { SpinningLoader } from '../../Loaders'
 import { Component as ComponentType, LinkState, Props, Sizes } from './Link.types'
