+++ packages/constellation/src/components/Primitives/Buttons/Link/Link.types.ts
@@ -1,5 +1,7 @@
 import React from 'react'
 
+import { PolymorphicComponentPropsWithRef } from '../../../../typings/polymorphic-component-props-with-ref'
+
 export type Sizes = 'default' | 'large' | 'small'
 export type LinkState = 'initial' | 'loading' | 'success'
