+++ packages/constellation/src/typings/polymorphic-ref.ts
@@ -1,0 +1,2 @@
+// Extract the `ref` prop type from a polymorphic component.
+export type PolymorphicRef<T extends React.ElementType> = React.ComponentPropsWithRef<T>['ref']
